fix: sio_timer timeout calculation overflow

+++ sio_timer.c
@@ -51,7 +51,7 @@ struct sio_timer *sio_timer_create(void *(*routine)(void *arg), void *arg)
     return timer;
 }
 
-int sio_timer_start(struct sio_timer *timer, unsigned int udelay, unsigned int uperiod)
+int sio_timer_start(struct sio_timer *timer, unsigned long int udelay, unsigned long int uperiod)
 {
     SIO_COND_CHECK_RETURN_VAL(!timer, -1);
